\name{Bordeaux}
\docType{data}
\alias{Bordeaux}
\title{Ratings of Bordeaux Wines}
\description{
A dataset comparing the rating of expensive bordeaux wines from two prominent wine critics.
}
\usage{Bordeaux}
\format{A dataframe with 72 observations.}
\source{Compiled by Dr. Simon Sheather.}
\keyword{datasets}