\name{MichelinNY}
\docType{data}
\alias{MichelinNY}
\title{Data from the Michelin Guide for New York City Resaurants}
\description{
The Michelin Guide publishes travel information world wide.  An award that the Michelin Guide give is the coveted
michelin star, a star in the michelin resaurant guide.  The awards are given based on a confidential evaluation process.
The Zagat Survey on the other hand is reviews and ratings based on comsumer input.\\
The dataset MichelinNY contains information from both the Michelin Guide and the Zagat Survey.
}
\usage{data(MichelinNY)}
\format{A dataframe with 165 observations.}
\source{
\emph{Michelin Guide New York City 2006}, Michelin Travel Publications, Greenville, South Carolina\\
Zagat survey:2006 New York City Restaurants, Available online at www.zagat.com
}
\keyword{datasets}