\name{MissAmerica08}
\docType{data}
\alias{MissAmerica08}
\title{Data for finalists in the Miss America Pageant}
\description{
A dataset for the miss america pageant for the years of 2000 through 2008.  
Included demographic and geographic data for each of the 50 states plus DC.
}
\usage{data(MissAmerica08)}
\format{A dataframe with 51 observations.}
\source{Compiled by Dr. Simon Sheather.}
\keyword{datasets}